[Refactor] Move scan range filter primitives to storage primitive
