+++ be/src/exec/olap_scan_prepare.h
@@ -19,8 +19,8 @@
 #include "exec/olap_common.h"
 #include "exprs/expr.h"
 #include "exprs/expr_context.h"
-#include "filter_condition.h"
 #include "runtime/descriptors.h"
+#include "storage/primitive/filter_condition.h"
 #include "storage/primitive/predicate_tree/predicate_tree_fwd.h"
 #include "storage/primitive/predicate_tree_params.h"
 #include "storage/runtime_filter_predicate.h"
