+++ be/src/exec/pipeline/set/except_context.cpp
@@ -14,7 +14,9 @@
 
 #include "exec/pipeline/set/except_context.h"
 
+#include "exprs/expr.h"
 #include "runtime/current_thread.h"
+#include "runtime/descriptors.h"
 #include "runtime/exec_env.h"
 #include "runtime/runtime_state.h"
