+++ be/src/exec/pipeline/set/except_context.h
@@ -23,11 +23,11 @@
 #include "column/runtime_type_traits.h"
 #include "common/statusor.h"
 #include "exec/except_hash_set.h"
-#include "exec/olap_common.h"
 #include "exec/pipeline/context_with_dependency.h"
 #include "exec/pipeline/primitives/pipeline_observer.h"
 #include "exprs/expr_context.h"
 #include "gutil/casts.h"
+#include "runtime/descriptors_fwd.h"
 #include "runtime/mem_pool.h"
 
 namespace starrocks::pipeline {
