+++ be/src/storage/rowset/segment_iterator.cpp
@@ -47,7 +47,6 @@
 #include "segment_options.h"
 #include "storage/chunk_helper.h"
 #include "storage/column_expr_predicate.h"
-#include "storage/column_or_predicate.h"
 #include "storage/column_predicate_factory.h"
 #include "storage/column_predicate_inverted_index_fallback.h"
 #include "storage/column_predicate_rewriter.h"
@@ -60,6 +59,7 @@
 #include "storage/lake/filenames.h"
 #include "storage/lake/update_manager.h"
 #include "storage/primitive/chunk_iterator.h"
+#include "storage/primitive/column_or_predicate.h"
 #include "storage/primitive/projection_iterator.h"
 #include "storage/primitive/range.h"
 #include "storage/primitive/roaring2range.h"
