+++ be/test/storage/column_and_predicate_test.cpp
@@ -12,11 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "storage/column_and_predicate.h"
+#include "storage/primitive/column_and_predicate.h"
 
 #include <gtest/gtest.h>
 
 #include "common/block_split_bloom_filter.h"
+#include "storage/column_predicate_factory.h"
 #include "testutil/column_test_helper.h"
 
 namespace starrocks {
